Metabolite search: some code consolidation
